Add missing benchmarks (#11)
* Refactor *stacktraceTree.insert benchmark to compare default sizes

The goal of this change is to have data to come up with a better
default for `defaultStacktraceTreeSize` (currently 0). Analyzing these
results with `benchstat` shows improvements already in allocations and
bytes per operation with lower default sizes, an very big savings in
CPU and memory for a default big sized tree.

* Add *PartitionWriter.WriteProfileSymbols benchmark

* enable differential benchmark workflow for pkg/phlaredb

---------

Co-authored-by: Tabor Henderson <tabor.henderson@grafana.com>